Fixing sacss failure introduced in changelist 183045.  alignments are unpacked from varargs as promoted to size_t, but mps_pf_align is a naked number literal, and so was passed as int, causing a corrupt alignemnt failure in sacss.
Copied from Perforce
 Change: 183048
 ServerID: perforce.ravenbrook.com

+++ mps/code/apss.c
@@ -158,7 +158,7 @@ static void testInArena(mps_arena_t arena, mps_pool_debug_option_s *options)
   /* yet (MV Debug works here, because it fakes it through PoolAlloc). */
   printf("MVFF\n");
   res = stress(mps_class_mvff(), randomSizeAligned, arena,
-               (size_t)65536, (size_t)32, MPS_PF_ALIGN, TRUE, TRUE, TRUE);
+               (size_t)65536, (size_t)32, (size_t)MPS_PF_ALIGN, TRUE, TRUE, TRUE);
   if (res == MPS_RES_COMMIT_LIMIT) return;
   die(res, "stress MVFF");
